% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m49_countries}
\alias{m49_countries}
\title{M49 countries as a relational table table.}
\format{An object of class \code{data.frame} with 249 rows and 9 columns.}
\usage{
m49_countries
}
\description{
M49 countries as a relational table table.
}
\keyword{datasets}
