% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m49_full}
\alias{m49_full}
\title{M49 data as a flat table.}
\format{An object of class \code{data.frame} with 249 rows and 16 columns.}
\usage{
m49_full
}
\description{
M49 data as a flat table.
}
\keyword{datasets}
