% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m49_regions}
\alias{m49_regions}
\title{M49 regions as a relational table table.}
\format{An object of class \code{data.frame} with 31 rows and 3 columns.}
\usage{
m49_regions
}
\description{
M49 regions as a relational table table.
}
\keyword{datasets}
